Modify l2 firmware to be k1c_firmware
Summary:
This package will also contains ethernet firmware to be installed in
target. At the moment, only the l2_cache binary is present and will be
used by post-image script.
The idea is to deliver a package that contains firmware to be used with
linux.

Test Plan: Build buildroot completely to check that everything is building.

Reviewers: O53 buildroot!, O51 linux!

Differential Revision: https://phab.kalray.eu/D764