longvalid: validate test status
Summary:
In order to be sure we have no regression, add expected status for each
tests. This will be matched in check_vmlinux.sh to verify that all tests
are correctly passed/fail according to the expected status.
This will eventually be picked up in some junit reporting.

Test Plan:
Ran the lognvalid and put and unexpected status for one test. Testsuite
failed as expected.

Reviewers: O53 Buildroot Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D1284