Move uclibc-ng test in common configuration
Summary:
As now uclibc-ng test are executed from Robotframework we move it in the
common configuration. The shuffling and the run of uclibcng test
framework is now managed by Robot so pre_images.sh and scripts for
uclibc-ng tests have been removed.

Depends on D2042

Test Plan: Validate uclibc-ng on hardware if SMP is enabled

Reviewers: O53 Buildroot Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D2067