local.mk: Use path as described in the documentation
Summary:
In order to allow the user rebuilding an image using the documentation
and with as few modification as possible, add path as described in
documentation. If the user follows the userguide, it will be able to
rebuild an image easily.

Depends On D2883

Test Plan: Rebuild buildroot using new configs

Reviewers: O53 Buildroot Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D2884