Package: add mppa_rproc package.
Summary:
This commit add mppa_rproc package. This package uses mppa_rproc sources
which are provided by an extra clone. Not that since this package uses
the kalray Makefile, the kalray makefile and modules are copied inside
support/kalray/ in br2_external to allow using it from provided package.
Then to build it, the .mk file export K1_TOOLCHAIN_DIR to this path to
allow using the kalray makefile seamlessly.

Ref T11807

Test Plan: A test will be introduced in test_src in valid

Reviewers: O53 Buildroot Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T11807

Differential Revision: https://phab.kalray.eu/D3657