Enable stress-ng
Summary:
This patch enables the build of stress-ng and adds a patch to put the
nop assembly in a bundle. Without this there are too many instructions
in bundle.

Ref T11003

Test Plan:
stress-ng will probably be used for daily or weekly. For the
moment just ensure that it is build.

Reviewers: O53 Buildroot Coolidge, ysionneau, cleger

Reviewed By: O53 Buildroot Coolidge, ysionneau, cleger

Subscribers: cleger, #linux_coolidge_cc

Maniphest Tasks: T11003

Differential Revision: https://phab.kalray.eu/D2235