packages/pocl: remove dependency on firmware and add default firmware
Summary:
In order to make it possible to build OpenCL application with the
prebuitl toolchain but without bloating the final image, disable the
dependency between pocl and pocl-firmware. pocl build will be enabled in
the standard image and a default firmware will be included allowing
to build and run an OpenCL application.
POCL_FIRMWARE configuration has also been modified to select only
a set of firmwares to be installed.

Ref T13347

Test Plan:
Compile image and verify that opencl tests are now buildable and
runnable.

Reviewers: O53 Buildroot Coolidge, gthouvenin

Reviewed By: O53 Buildroot Coolidge, gthouvenin

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T13347

Differential Revision: https://phab.kalray.eu/D6465