rescue: enable link-local usage through CDC ECM
Summary:
The rescue configuration is designed to be smaller, in order to fit into
SPI NOR, for example. But not being able to use the link local through
the USB gadget is rather limiting.

This patch activates the required tools to enable ssh to the rescue
conf.

Test Plan: RAUC PoC test

Reviewers: O53 Buildroot Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D13057