rootfs: add a rauc config for cooperative boot
Summary:
The RAUC PoC used the default RAUC settings, but since we now have to
add support for a new bootloader type, we have to give it a custom
config.

This config has been added to all images for now. If there is a need
somewhere to use RAUC without kvx cooperative boot, it may be moved to a
specific buildroot config.

Test Plan: RAUC PoC in weekly

Reviewers: O53 Buildroot Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D14399