rauc: change generated buildroot package name
Summary:
The buildroot package generated was actually partially conflicting with
the existing buildroot rauc package, which was inducing strange build
behaviours and constrained the package version, for example.

This patch duplicates the package with another name in order to get a
better handle on it.

It comes with built-in tarball support.

Test Plan:
Check that all parameters can be changed without BuildRoot complaining or
splurting strange errors...

Reviewers: O53 Buildroot Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D14814