init.d: usbgadget: Warn early if no UDC found
Summary:
If there are no USB Device Controller (UDC) then the diretory
`/sys/class/udc/` is empty and the command:
    ls /sys/class/udc/ | head -n1 > UDC
will not enable the usb gadget on the first available controller,
but it will silently disable the gadget (by writing en empty string).

Add a test early in script that will warn and exit with error if there
are no UDC found.

ref T19685

Test Plan: CI

Reviewers: O53 Buildroot Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D22182