init.d: usbgadget: Fix write to configfs
Summary:
The init script that configure the usbgadget fails when using musl.
The script fails to write configuration values into configfs files.
For instance, the following two lines are generating errors:
    echo "0x0200" > /sys/kernel/config/usb_gadget/mppalinux/bcdUSB
    echo "20080000.usb0" > /sys/kernel/config/usb_gadget/mppalinux/UDC

The first line returns the following error:
    echo: write error: Invalid argument

The second line does not behave correctly and no gadget is started,
with strace we can see the following system call:
    writev(1, [{iov_base="20080000.usb0", iov_len=13}, "\n", 2)
The writev system call is generated by musl's implementation of printf.

The Linux kernel treat writev as two different writes and this is *very*
problematic when writting to /sys/* files: It is treated as two distinct
write. When writing in configfs we really want only one write with the
full configuration value.

There is no guaranties on which syscall is used when using stdio: any
functions that operates on FILE. The portable way to fix this issue is
to use a program that specifically uses the write system call (and
hopefully only one). The simplest way to achieve this is by using the
command dd.

ref T19661

Test Plan: CI

Reviewers: O53 Buildroot Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D22183