config: add github config
Summary:
Add github config

This allows GitHub users to build our Linux
kvx.config is not usable since it also selects
packages that are not published like nvmem.
It also tries to add l2$ firmware which is not published.

github config is based on kvx.config without what's not
published.

kvx.config was needed in the first place because it selects
BR2_PACKAGE_HOST_LINUX_HEADERS_CUSTOM_5_10=y
Without this, by default 5_13 is selected which fails when
buildroot finds out our Linux kernel really is 5.10
and not 5.13+.

Ref T14925

Test Plan: CI

Reviewers: O53 Buildroot Coolidge, jvetter

Reviewed By: O53 Buildroot Coolidge, jvetter

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D22283