Packages now use a tarball by default instead of the source dir
Summary:
Now by default Buildroot uses tarballs for all Kalray packages, and for the
tests in the integration we point Buildroot to our sources via a toplevel
`local.mk`. This is the intended way to work with buildroot. This
removed the need to have for every package a `_CUSTOM_TARBALL` target. Also
otherwise a `local.mk` would need to be delivered to the customer. This might
lead to conflicts if they are already using an other `local.mk` for their
local packages. Only a single `local.mk` can be used, so a local merge would
be necessary which is dirty.

Ref T19916

Test Plan: CI

Reviewers: ysionneau, fpiedallu, O53 Buildroot Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T19916

Differential Revision: https://phab.kalray.eu/D23260