remove l2 firmware from the packages
Summary:
L2 firmware should not be necessary anymore in the LTC repository, as it
is handled by lower-level boot stages.

This patch removes the packaging of the L2 firmware and configs
referring to it.

Ref T20630

Test Plan:
Build of l2 should now be skipped, and post_image should not need to
generate a vmlinux with .rm_firmware section patched

Reviewers: O53 Buildroot Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T20630

Differential Revision: https://phab.kalray.eu/D26275