config: Add BR2_TOOLCHAIN_GCC_AT_LEAST_12 option
Summary:
Buildroot requires a specific gcc version to build some packages (gdb
for example). When we set a custom version with BR2_GCC_CUSTOM_VERSION
like in kvx_defconfig, buildroot is unable to determine the version of
gcc so some packages are not built.
Ideally, we should dynamically derive the version of gcc from the sha1
we put in BR2_GCC_CUSTOM_VERSION and add the right
BR2_TOOLCHAIN_GCC_AT_LEAST_X but I couldn't find an easy way to do that.
Ref T21853

Test Plan: NA

Reviewers: jvetter, O53 Buildroot Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T21853

Differential Revision: https://phab.kalray.eu/D32158